Mode functions to utilities & add option to calibrate sigma on the fly
